(PC-14539)[API] fix: serialize booking amount as float to display decimal

+++ api/src/pcapi/routes/serialization/bookings_recap_serialize.py
@@ -51,7 +51,7 @@ class BookingRecapResponseBookingStatusHistoryModel(BaseModel):
 
 class BookingRecapResponseModel(BaseModel):
     beneficiary: BookingRecapResponseBeneficiaryModel
-    booking_amount: int
+    booking_amount: float
     booking_date: datetime
     booking_is_duo: bool
     booking_status: BookingRecapStatus
